move toxiproxy fixtures to conftest
